📚 Ignore/exclude `.ipynb` files
One of the main reasons for using MyST Markdown is to have all the documentation content
in Markdown format instead of the JSON format of `.ipynb` files, which is hard to parse
for a human and results in messy diffs. My current setup for writing the documentation
is to write a `.ipynb` file and pair it to a `.md` equivalent using Jupytext and a
VSCode extension. So it's useful to have the `.ipynb` files in the documentation
directories during writing, but I don't want to commit them to the repository.

In any case, we should try to avoid `.ipynb` files in the documentation in favor of
`.md` ones. So here we add a `.gitignore` to do so, as well as exclude them from the
MyST documentation build.