✨ Add the `EnergyLevelsWorkChain`
Add a first draft of the `EnergyLevelsWorkChain`, to calculate the energy levels of a
defect structure. The work chain logic will be documented in more detail once we
converge on a final version. We have also added a default protocol to make it much
easier to run the workflow.

Also add some basic documentation to set up the computers/codes to
run the workflow, for now still _very_ specific to Pawsey.