💥 `EnergyLevelsWorkChain`: remove TDM logic
Calculating the transition dipole moments (TDM) via the VASPKIT script it unnecessarily
convoluted and fragile. There are other tools out there for parsing the WAVECAR file,
and calculating the TDM by hand is relatively simple.

Hence, we keep the focus of the `EnergyLevelsWorkChain` on optimizing the structure and
calculating the energy levels, removing any logic related to the TDM.